사용자 반환 타입 통일 및 직급별 수당 관리 기능 추가

+++ src/main/java/kr/ac/deu/computer_engineering/Absenteeism/Management/route/allowance/AllowanceController.java
@@ -4,18 +4,24 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.AllowanceOfRank.AllowanceOfRank;
 import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.AllowanceOfRank.dto.AllowanceOfRankDto;
-import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.AllowanceOfRank.dto.AllowanceOfRankDto;
+import kr.ac.deu.computer_engineering.Absenteeism.Management.enums.ResState;
+import kr.ac.deu.computer_engineering.Absenteeism.Management.handler.exception.ResponseDTO;
 import kr.ac.deu.computer_engineering.Absenteeism.Management.service.allowance.AllowanceService;
+import kr.ac.deu.computer_engineering.Absenteeism.Management.utils.RoleValidate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+import javax.validation.Valid;
 import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping(value = "/api/allowance")
+@RequestMapping(value = "/api/rank")
 public class AllowanceController {
     final private AllowanceService allowanceService;
 
@@ -23,52 +29,83 @@ public class AllowanceController {
     @Operation(
             summary = "직급별 수당 목록 조회",
             description = "직급별 수당 목록 조회")
-    @GetMapping("")
-    public ResponseEntity<?> getItemList(@RequestParam(required = true) Long rankId) {
-        List<AllowanceOfRank> result = allowanceService.getList(rankId);
-        return new ResponseEntity<>(result, HttpStatus.OK);
+    @GetMapping("/{rankId}/allowance")
+    public ResponseEntity<?> getItemList(
+            HttpServletRequest request,
+            @PathVariable Long rankId) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session)) {
+            List<AllowanceOfRank> result = allowanceService.getList(rankId);
+            return new ResponseEntity<>(new ResponseDTO<>(ResState.OK, result), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @Tag(name = "직급별 수당")
     @Operation(
             summary = "직급별 수당 Id로 상세 조회",
             description = "직급별 수당 상세 조회")
-    @GetMapping("/{id}")
-    public ResponseEntity<?> getItemById(@PathVariable Long id) {
-        AllowanceOfRank allowanceOfRank = allowanceService.getAccountById(id);
-        return new ResponseEntity<>(allowanceOfRank, HttpStatus.OK);
+    @GetMapping("/{rankId}/allowance/{id}")
+    public ResponseEntity<?> getItemById(
+            HttpServletRequest request,
+            @PathVariable Long rankId,
+            @PathVariable Long id) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session)) {
+            AllowanceOfRank result = allowanceService.getAccountByRankIdAndId(rankId, id);
+            return new ResponseEntity<>(new ResponseDTO<>(ResState.OK, result), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @Tag(name = "직급별 수당")
     @Operation(
             summary = "직급별 수당 정보 생성",
             description = "직급별 수당 정보 생성")
-    @PostMapping("")
-    public ResponseEntity<?> createItem(@RequestBody AllowanceOfRankDto dto) {
-        allowanceService.createAllowanceOfRank(dto);
-        return new ResponseEntity<>(HttpStatus.CREATED);
+    @PostMapping("/allowance")
+    public ResponseEntity<?> createItem(
+            HttpServletRequest request,
+            @Valid @RequestBody AllowanceOfRankDto dto,
+            BindingResult bindingResult) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session)) {
+            allowanceService.createAllowanceOfRank(dto);
+            return new ResponseEntity<>(new ResponseDTO<>(ResState.CREATED), HttpStatus.CREATED);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @Tag(name = "직급별 수당")
     @Operation(
             summary = "직급별 수당 Id로 직급별 수당 정보 수정",
             description = "직급별 수당 정보 수정")
-    @PutMapping("/{id}")
+    @PutMapping("/allowance/{id}")
     public ResponseEntity<?> updateItemById(
+            HttpServletRequest request,
             @PathVariable Long id,
-            @RequestBody AllowanceOfRankDto dto
-        ) {
-        allowanceService.updateAllowanceOfRank(id, dto);
-        return new ResponseEntity<>(HttpStatus.OK);
+            @Valid @RequestBody AllowanceOfRankDto dto,
+            BindingResult bindingResult) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session)) {
+            allowanceService.updateAllowanceOfRank(id, dto);
+            return new ResponseEntity<>(new ResponseDTO<>(ResState.OK), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @Tag(name = "직급별 수당")
     @Operation(
             summary = "직급별 수당 Id로 직급별 수당 정보 삭제",
             description = "직급별 수당 정보 삭제")
-    @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteItemById(@PathVariable Long id) {
-        allowanceService.deleteAllowanceOfRank(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+    @DeleteMapping("/allowance/{id}")
+    public ResponseEntity<?> deleteItemById(
+            HttpServletRequest request,
+            @PathVariable Long id) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session)) {
+            allowanceService.deleteAllowanceOfRank(id);
+            return new ResponseEntity<>(new ResponseDTO<>(ResState.OK), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 }
